+++ src/main/java/roomescape/reservation/controller/AdminReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.admin;
+package roomescape.reservation.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
